requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.6.1
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckdb>=0.9.0
scipy>=1.11.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiofiles>=0.23.0
python-dotenv>=1.0.0
Pillow>=9.0.0